Add expressions and conditions withdrawal
What? Now, if an AND condition consists of only
one REL condition, it will 'withdraw itself'
and return as parsing result only the REL condition.

Why? This pattern avoids unnecessary structure
wrapping, and makes returned result cleaner.
If fact, OR condition with only single AND
condition is in reality AND condition itself.